+++ db/migrate/20111002060240_create_gears.rb
@@ -2,9 +2,9 @@ class CreateGears < ActiveRecord::Migration
   def self.up
     create_table :gears do |t|
       t.string :name, :null => false
-      t.text :description, :null => true
-      t.double :price, :null => false
-      t.double :weight, :null => false
+      t.text   :description, :null => true
+      t.float  :price, :null => false
+      t.float  :weight, :null => false
     end
   end
